refactor[#231] : 마킹 상수를 import하는 경로 수정
