+++ src/features/marking/api/search.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { useResearchMarkingList } from "@/features/map/hooks";
 import { useAuthStore } from "@/shared/store";
-import { MARKING_REQUEST_URL } from "../contants/requestUrl";
+import { MARKING_REQUEST_URL } from "../constants/requestUrl";
 
 interface Pet {
   petId: number;
